requirements.txt
fastapi
uvicorn
injector
python-dotenv
requests
beautifulsoup4
click
pytest
httpx
playwright
camoufox
